+++ src/owoify_clj/utility/presets.clj
@@ -5,14 +5,21 @@
                                  mappings/map-mom-to-mwom
                                  mappings/map-time-to-tim
                                  mappings/map-me-to-mwe
-                                 mappings/map-n-vowel-to-ny
                                  mappings/map-over-to-owor
                                  mappings/map-ove-to-uv
                                  mappings/map-haha-to-hehe-xd
                                  mappings/map-the-to-teh
                                  mappings/map-you-to-u
                                  mappings/map-read-to-wead
-                                 mappings/map-worse-to-wose])
+                                 mappings/map-worse-to-wose
+                                 mappings/map-great-to-gwate
+                                 mappings/map-aviat-to-awiat
+                                 mappings/map-dedicat-to-deditat
+                                 mappings/map-remember-to-rember
+                                 mappings/map-when-to-wen
+                                 mappings/map-frightened-to-frigten
+                                 mappings/map-meme-to-mem
+                                 mappings/map-feel-to-fell])
 
 (def uvu-mapping-list [mappings/map-o-to-owo
                        mappings/map-ew-to-uwu
@@ -29,7 +36,8 @@
                        mappings/map-ry-to-wwy
                        mappings/map-r-or-l-to-w])
 
-(def owo-mappings-list [mappings/map-ll-to-ww
+(def owo-mappings-list [mappings/map-n-vowel-to-ny
+                        mappings/map-ll-to-ww
                         mappings/map-vowel-or-r-except-o-l-to-wl
                         mappings/map-old-to-owld
                         mappings/map-ol-to-owl
@@ -43,4 +51,6 @@
                         mappings/map-consonant-r-to-consonant-w
                         mappings/map-ly-to-wy
                         mappings/map-ple-to-pwe
-                        mappings/map-nr-to-nw])
+                        mappings/map-nr-to-nw
+                        mappings/map-mem-to-mwem
+                        mappings/unmap-nywo-to-nyo])