+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## 1.0.0 - 2023-12-05
+### Changed
+- Some mappings in some presets. The same input string probably won't give the same result as version 0.1.x.
+### Added
+- The `uwuify` and `uvuify` function for convenience.
+
 ## 0.1.1 - 2021-06-11
 ### Changed
 - Update README.md